split lsp input validation into dedicated microcrate
