+++ crates/perl-lsp-input-validation/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "perl-lsp-input-validation"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "LSP request and file-input validation helpers"
+documentation = "https://docs.rs/perl-lsp-input-validation"
+readme = "README.md"
+keywords = ["perl", "lsp", "validation", "security"]
+categories = ["development-tools", "text-editors"]
+include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
+
+[dependencies]
+anyhow = "1.0.102"
+serde_json = { workspace = true }
+perl-path-security = { workspace = true }
+
+[dev-dependencies]
+perl-tdd-support = { workspace = true }
+tempfile = "3.24.0"
+
+[lints]
+workspace = true
