+++ crates/perl-lsp-input-validation/README.md
@@ -1,0 +1,10 @@
+# perl-lsp-input-validation
+
+Standalone microcrate for LSP request and source-input validation.
+
+## Scope
+
+- Workspace-bound file path validation with Perl extension filtering.
+- Text/document payload size and content checks.
+- LSP method/parameter validation helpers.
+- String sanitization and workspace-root safety checks.
