+++ crates/perl-lsp-input-validation/ROADMAP.md
@@ -1,0 +1,10 @@
+# perl-lsp-input-validation Roadmap
+
+> **Note:** This is the component-specific roadmap for `perl-lsp-input-validation`. For the project-wide roadmap, see [`docs/ROADMAP.md`](../../docs/ROADMAP.md).
+
+## Purpose
+LSP request and file-input validation helpers.
+
+## Current Status (v0.10.0)
+- **Status:** Initial Public Alpha
+- **Integration:** Shared by `perl-lsp`.
